requirements.txt
# API
fastapi
uvicorn[standard]

# Database
sqlalchemy
psycopg2-binary

# ML & Data
pandas
scikit-learn
scikit-surprise  # For recommendation
mlflow
joblib

# Testing
pytest
httpx # For TestClient